[FEAT] 회사 정보 조회 연결 완료
