+++ back/src/main/java/com/thered/stocksignal/service/company/CompanyService.java
@@ -1,8 +1,6 @@
 package com.thered.stocksignal.service.company;
 
-import com.thered.stocksignal.domain.entity.Company;
-
-import java.util.Optional;
+import static com.thered.stocksignal.app.dto.CompanyDto.*;
 
 public interface CompanyService {
 
@@ -11,5 +9,8 @@ public interface CompanyService {
 
     // 회사명 -> 로고 조회
     String findLogoByName(String companyName);
+
+    // 종목 코드 -> 회사 정보 (분석 탭에 들어갈 내용)
+    CompanyInfoResponseDto findCompanyInfoByCode(String companyCode, String accessToken, String appKey, String appSecret);
 }
